rename generator dir to builder
