+++ LICENSE
@@ -1,0 +1,17 @@
+Copyright (c) 2025 James Reeves
+
+All rights reserved.
+
+Permission is hereby granted to any person obtaining a copy of this software and associated documentation files (the "Software"), to use the Software for personal, educational, or internal business purposes only, subject to the following conditions:
+
+1. The Software may not be modified, copied, reproduced, published, distributed, sublicensed, sold, or otherwise exploited for any commercial purpose without the express prior written consent of the copyright holder.
+
+2. No derivative works based on the Software may be created without express prior written consent.
+
+3. The Software is provided "as is", without warranty of any kind, express or implied, including but not limited to the warranties of merchantability, fitness for a particular purpose, and noninfringement.
+
+4. Unauthorized use, copying, reproduction, modification, distribution, or derivative works are strictly prohibited and may result in civil and/or criminal penalties.
+
+If you wish to license the Software for any other purpose, you must obtain explicit written permission from the copyright holder.
+
+Contact: james@codebrojim.com
